replace "verion.Short()" to "version.Long()"

+++ cmd/ingress/ingress.go
@@ -97,7 +97,7 @@ the apisix cluster and others are created`,
 			log.DefaultLogger = logger
 			log.Info("apisix ingress controller started")
 
-			log.Info("version: ", version.Short())
+			log.Info("version: ", version.Long())
 
 			data, err := json.MarshalIndent(cfg, "", "\t")
 			if err != nil {
